requirements.txt
# P21 API Documentation - Dependencies

# HTTP client (sync and async)
httpx>=0.25.0

# Environment management
python-dotenv>=1.0.0

# Markdown to HTML conversion
markdown>=3.5.0
